retry: allow immutable options to be passed in.
The options to retry may come from a config system. config systems like
node-config make their configuration immutable for safety.

However, the currently logic of retry was modifying the input by
reference, causing an exception to be thrown.

I don't expect it was the intent to modify configuration values passed
in, which may be re-used elsewhere.

This patch updates the code to clone the options passed in so we have
a mutable copy of the input.

Note that Object.assign() does not throw if the input is null or
undefined, so it's no longer necessary to check if the options passed in
were  an object already.

Ref: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/assign